Allow images to be distributed as container images
Containerfile.image by default uses centos:stream9-minimal as a base
and packages edpm-hardened-uefi.qcow2 (all overridable by buildah bud
--build-arg). Containerfile.ramdisk does likewise for
ironic-python-agent.

The README.rst demonstrates copying the image files out of a container
to a local directory, but the real advantage of this approach is
running this as an init container in a pod to populate a persistent
volume with images that can then be served over http by a pod service.