Enable FIPS for just the RHEL image
FIPS is currently not enabled on the CentOS image because upstream CI
cannot consistently provide enough entropy to pass FIPS requirements.
However FIPS still needs to be enabled by default, so this change adds
FIPS support to the RHEL EDPM image.

Similar issues may be found downstream but we need to try it to find
out. The workaround may be different downstream, such as getting VM
based jobs to modify the image to disable FIPS on-the-fly.

Issue: OSPRH-1196