Replace efi grub.cfg with RHEL style redirect
Currently the efi grub.cfg is identical to the /boot/grub2/grub.cfg,
which only occurs thanks to a second manual call to grub2-mkconfig. This
risks out-of-band updates resulting in a stale efi grub.cfg, which in
some cases may prevent the host from booting at all.

The RHEL efi grub.cfg is a simple redirect to the /boot partition. This
change replicates this redirect.

A corresponding change to edpm-ansible will stop overwriting efi
grub.cfg with the previous approach.